move legacysmartmodule payload to spu_schema to decouple SmartEngine
