+++ crates/fluvio-smartengine/src/error.rs
