+++ crates/fluvio-smartengine/src/lib.rs
@@ -1,6 +1,9 @@
+
+pub(crate) mod error;
+pub(crate) mod memory;
+
+pub mod transforms;
+mod engine;
+pub use engine::*;
+
 pub type WasmSlice = (i32, i32, u32);
-
-#[cfg(feature = "engine")]
-pub mod engine;
-
-pub mod metadata;
