+++ crates/fluvio-smartengine/src/memory.rs
