+++ crates/fluvio-smartengine/src/transforms/aggregate.rs
@@ -9,9 +9,9 @@ use fluvio_smartmodule::dataplane::smartmodule::{
     SmartModuleExtraParams, SmartModuleInput, SmartModuleOutput, SmartModuleAggregateInput,
     SmartModuleInternalError, SmartModuleAggregateOutput,
 };
-use crate::WasmSlice;
+use crate::{WasmSlice, SmartModuleWithEngine, SmartModuleContext, error::Error, SmartModuleInstance};
 
-use super::{SmartModuleContext, SmartModuleWithEngine, error::Error, SmartModuleInstance};
+
 
 const AGGREGATE_FN_NAME: &str = "aggregate";
 type OldAggregateFn = TypedFunc<(i32, i32), i32>;
