+++ crates/fluvio-smartengine/src/transforms/map.rs
@@ -5,73 +5,72 @@ use anyhow::Result;
 use fluvio_smartmodule::dataplane::smartmodule::{
     SmartModuleExtraParams, SmartModuleInput, SmartModuleOutput, SmartModuleInternalError,
 };
-use tracing::{debug, instrument};
+use tracing::debug;
 use wasmtime::{AsContextMut, Trap, TypedFunc};
 
-use crate::WasmSlice;
+use crate::{WasmSlice, SmartModuleWithEngine, SmartModuleContext, error::Error, SmartModuleInstance};
 
-use super::{SmartModuleContext, SmartModuleWithEngine, error::Error, SmartModuleInstance};
-
-const JOIN_FN_NAME: &str = "join";
-type OldJoinFn = TypedFunc<(i32, i32), i32>;
-type JoinFn = TypedFunc<(i32, i32, u32), i32>;
+const MAP_FN_NAME: &str = "map";
+type OldMapFn = TypedFunc<(i32, i32), i32>;
+type MapFn = TypedFunc<(i32, i32, u32), i32>;
 
 #[derive(Debug)]
-pub struct SmartModuleJoin {
+pub struct SmartModuleMap {
     base: SmartModuleContext,
-    join_fn: JoinFnKind,
+    map_fn: MapFnKind,
+}
+enum MapFnKind {
+    Old(OldMapFn),
+    New(MapFn),
 }
 
-pub enum JoinFnKind {
-    Old(OldJoinFn),
-    New(JoinFn),
-}
-
-impl Debug for JoinFnKind {
+impl Debug for MapFnKind {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Old(_join_fn) => write!(f, "OldJoinFn"),
-            Self::New(_join_fn) => write!(f, "JoinFn"),
+            Self::Old(..) => write!(f, "OldMapFn"),
+            Self::New(..) => write!(f, "MapFn"),
         }
     }
 }
 
-impl JoinFnKind {
+impl MapFnKind {
     fn call(&self, store: impl AsContextMut, slice: WasmSlice) -> Result<i32, Trap> {
         match self {
-            Self::Old(join_fn) => join_fn.call(store, (slice.0, slice.1)),
-            Self::New(join_fn) => join_fn.call(store, slice),
+            Self::Old(map_fn) => map_fn.call(store, (slice.0, slice.1)),
+            Self::New(map_fn) => map_fn.call(store, slice),
         }
     }
 }
 
-impl SmartModuleJoin {
+impl SmartModuleMap {
+    #[tracing::instrument(skip(module, params))]
     pub fn new(
         module: &SmartModuleWithEngine,
         params: SmartModuleExtraParams,
         version: i16,
     ) -> Result<Self, Error> {
+        debug!(base_fn = MAP_FN_NAME, ?params, "instantiating mapping");
         let mut base = SmartModuleContext::new(module, params, version)?;
-        let join_fn =
-            if let Ok(join_fn) = base.instance.get_typed_func(&mut base.store, JOIN_FN_NAME) {
-                JoinFnKind::New(join_fn)
-            } else {
-                let join_fn = base
-                    .instance
-                    .get_typed_func(&mut base.store, JOIN_FN_NAME)
-                    .map_err(|err| Error::NotNamedExport(JOIN_FN_NAME, err))?;
-                JoinFnKind::Old(join_fn)
-            };
-        Ok(Self { base, join_fn })
+        let map_fn = if let Ok(map_fn) = base.instance.get_typed_func(&mut base.store, MAP_FN_NAME)
+        {
+            debug!("found new map function");
+            MapFnKind::New(map_fn)
+        } else {
+            debug!("not found map function");
+            let map_fn: OldMapFn = base
+                .instance
+                .get_typed_func(&mut base.store, MAP_FN_NAME)
+                .map_err(|err| Error::NotNamedExport(MAP_FN_NAME, err))?;
+            MapFnKind::Old(map_fn)
+        };
+        Ok(Self { base, map_fn })
     }
 }
 
-impl SmartModuleInstance for SmartModuleJoin {
-    #[instrument(skip(self, input), name = "Join")]
+impl SmartModuleInstance for SmartModuleMap {
     fn process(&mut self, input: SmartModuleInput) -> Result<SmartModuleOutput> {
         let slice = self.base.write_input(&input)?;
-        debug!(len = slice.1, "WASM SLICE");
-        let map_output = self.join_fn.call(&mut self.base.store, slice)?;
+        let map_output = self.map_fn.call(&mut self.base.store, slice)?;
 
         if map_output < 0 {
             let internal_error = SmartModuleInternalError::try_from(map_output)
