+++ crates/fluvio-smartengine/src/transforms/mod.rs
@@ -6,9 +6,3 @@ pub mod aggregate;
 pub mod join;
 pub mod file_batch;
 pub mod join_stream;
-pub(crate) mod error;
-
-mod inner;
-pub use inner::*;
-
-mod memory;
