+++ crates/fluvio-spu-schema/Cargo.toml
@@ -22,6 +22,7 @@ bytes = "1.0.0"
 serde = { version = "1.0.103", features = ['derive'] }
 static_assertions = "1.1.0"
 educe = { version = "0.4.19", features = ["Debug"] }
+flate2 = { version = "1.0.20" }
 
 fluvio-future = { version = "0.4.0", optional = true }
 # Fluvio dependencies
@@ -30,7 +31,7 @@ fluvio-protocol = { path = "../fluvio-protocol", version = "0.8.0", features = [
     "record",
 ] }
 fluvio-smartmodule = { path = "../fluvio-smartmodule", version = "0.4.0", default-features = false }
-fluvio-smartengine = { path = "../fluvio-smartengine", version = "0.4.0", default-features = false }
+
 
 [dev-dependencies]
 futures-util = { version = "0.3.4" }
