+++ crates/fluvio-spu-schema/src/server/mod.rs
@@ -1,6 +1,9 @@
 mod api_key;
+
+
 #[cfg(feature = "file")]
 mod api;
+pub mod smartmodule;
 pub mod fetch_offset;
 pub mod stream_fetch;
 pub mod update_offset;
