+++ crates/fluvio-spu-schema/src/server/smartmodule.rs
