+++ crates/fluvio-spu/Cargo.toml
@@ -64,7 +64,7 @@ fluvio-future = { version = "0.4.0", features = [
     "openssl_tls",
     "zero_copy",
 ] }
-fluvio-smartengine = { path = "../fluvio-smartengine", features = ["engine"] }
+fluvio-smartengine = { path = "../fluvio-smartengine" }
 fluvio-smartmodule = { path = "../fluvio-smartmodule",  default-features = false }
 
 [dev-dependencies]
