+++ crates/fluvio/Cargo.toml
@@ -16,7 +16,7 @@ path = "src/lib.rs"
 [features]
 default = []
 admin = ["fluvio-sc-schema/use_serde"]
-smartengine = ["fluvio-smartengine/engine"]
+smartengine = ["fluvio-smartengine"]
 stats = ["sysinfo", "quantities", "hdrhistogram", "strum", "num-traits"]
 unstable = []
 
@@ -70,7 +70,7 @@ fluvio-protocol = { path = "../fluvio-protocol", version = "0.8", features = [
 ] }
 fluvio-compression = { version = "0.2", path = "../fluvio-compression" }
 fluvio-smartmodule = { path = "../fluvio-smartmodule", version = "0.4.0", default-features = false }
-fluvio-smartengine = { path = "../fluvio-smartengine", version = "0.4.0", default-features = false }
+fluvio-smartengine = { path = "../fluvio-smartengine", version = "0.4.0", default-features = false, optional = true}
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 dirs = "4.0.0"
